refactoring app structure
