+++ src/pages/home/HomeView.vue
